--- app/src/main/java/com/example/agora/data/model/Comment.kt
@@ -1,4 +1,0 @@
-package com.example.agora.data.model
-
-class Comment {
-}
