--- app/src/main/java/com/example/agora/data/model/Notification.kt
@@ -1,4 +1,0 @@
-package com.example.agora.data.model
-
-class Notification {
-}
