--- app/src/main/java/com/example/agora/data/model/User.kt
@@ -1,4 +1,0 @@
-package com.example.agora.data.model
-
-class User {
-}
