+++ app/src/main/java/com/example/agora/model/data/Comment.kt
@@ -1,0 +1,4 @@
+package com.example.agora.model.data
+
+class Comment {
+}
