+++ app/src/main/java/com/example/agora/model/data/Notification.kt
@@ -1,0 +1,4 @@
+package com.example.agora.model.data
+
+class Notification {
+}
