+++ app/src/main/java/com/example/agora/screens/dashboard/DashboardViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.agora.ui.home
+package com.example.agora.screens.dashboard
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class DashboardViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
