+++ app/src/main/java/com/example/agora/screens/notifications/NotificationsViewModel.kt
@@ -1,13 +1,13 @@
-package com.example.agora.ui.dashboard
+package com.example.agora.screens.notifications
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class NotificationsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
